requirements.txt
setuptools>=68.0.0
pyusb>=1.2.1
meson>=1.2.0
ffmpeg
cairosvg>=2.7.1
numpy>=1.25.2
intelhex==2.3.0
urllib3
gcovr>=7.2
Pillow>=10.0.0
